connector/server/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
yoyo-migrations==9.0.0
psycopg2-binary==2.9.9
